refactor: renaming goal_processor component -> goal resolver + standard_agent doc string update
