+++ README.md
@@ -73,9 +73,6 @@ JENTIC_API_KEY="your-jentic-api-key-here"
 OPENAI_API_KEY="your-openai-api-key-here"
 ANTHROPIC_API_KEY="your-anthropic-api-key-here"
 GEMINI_API_KEY="your-google-gemini-api-key-here"
-
-# Tool-Specific Secrets (add as needed)
-DISCORD_BOT_TOKEN="your-discord-bot-token-here"
 ```
 
 **Note:** An LLM provider key is essential for the agent to function. The `JENTIC_API_KEY` is required if you are using the default `JenticClient` tool provider.
@@ -207,6 +204,9 @@ This composition approach means you can:
 
 The key insight is that each component follows well-defined interfaces (`BaseLLM`, `BaseMemory`, `JustInTimeToolingBase`, etc.), so they can be combined in any configuration that makes sense for your use case.
 
+#### Advanced Capability: Goal Processing
+To support more complex interactions, the `StandardAgent` can be configured with an optional **Goal Processor**. This is a swappable component that preprocesses a user's raw goal before the main reasoning loop begins. The pre-built `ReWOOAgent`, for example, includes an `ImplicitGoalResolver` that uses conversation history to understand follow-up questions. In the future, this same extension point could be used to add security layers that detect malicious intent or to sanitize inputs before execution.
+
 
 ### Project Layout
 
@@ -275,6 +275,8 @@ The framework is designed to be modular. Here are some common extension points:
 | **New tool provider**              | Create a class that inherits from `JustInTimeToolingBase`, implement its methods, and pass it to your `StandardAgent`.                                                               |
 | **Persistent memory**              | Create a class that implements the `MutableMapping` interface (e.g., using Redis), and pass it to your `StandardAgent`.                                                              |
 | **New Planners, Executors, etc.**  | Create your own implementations of `Plan`, `ExecuteStep`, `Reflect`, or `SummarizeResult` to invent new reasoning capabilities, then compose them in a `SequentialReasoner`. |
+| **Pre-process or validate goals**  | Create a class that inherits from `BaseGoalProcessor` and pass it to `StandardAgent`. Use this to resolve conversational ambiguities, check for malicious intent, or sanitize inputs. |
+
 
 ## 🔮 Roadmap
