--- agents/goal_processor/__init__.py
@@ -1,4 +1,0 @@
-from agents.goal_processor.base import BaseGoalResolver
-from agents.goal_processor.implicit_goal_resolver import ImplicitGoalResolver
-
-__all__ = ["BaseGoalResolver", "ImplicitGoalResolver"]
