+++ agents/goal_resolver/__init__.py
@@ -1,0 +1,4 @@
+from agents.goal_resolver.base import BaseGoalResolver
+from agents.goal_resolver.implicit_goal_resolver import ImplicitGoalResolver
+
+__all__ = ["BaseGoalResolver", "ImplicitGoalResolver"]
