+++ agents/goal_resolver/base.py
