+++ agents/prebuilt.py
@@ -3,7 +3,7 @@ from agents.tools.jentic import JenticClient
 from agents.memory.dict_memory import DictMemory
 from agents.reasoner.prebuilt import ReWOOReasoner
 from agents.llm.litellm import LiteLLM
-from agents.goal_processor.implicit_goal_resolver import ImplicitGoalResolver
+from agents.goal_resolver.implicit_goal_resolver import ImplicitGoalResolver
 
 
 class ReWOOAgent(StandardAgent):
@@ -15,6 +15,7 @@ class ReWOOAgent(StandardAgent):
     - JenticClient for external tool access
     - DictMemory for state persistence
     - ReWOO sequential reasoner for planning, execution, and reflection
+    - An ImplicitGoalResolver to enable conversational follow-ups.
     """
 
     def __init__(self, *, model: str | None = None, max_retries: int = 2):
